Rename test files to end with .test.js
