+++ package.json
@@ -12,7 +12,7 @@
     "create": "webpack",
     "create:dev": "webpack-dev-server --open",
     "build": "babel --presets es2015 app/template/js/src -d app/template/js/dist",
-    "test": "NODE_ENV=test ./node_modules/.bin/mocha test/*.js",
+    "test": "NODE_ENV=test ./node_modules/.bin/mocha test/*test.js",
     "migrations": "NODE_ENV=test sequelize db:migrate",
     "coverage": "node node_modules/.bin/istanbul cover _mocha tests/*tests.js",
     "coveralls": "cat ./coverage/lcov.info | node node_modules/.bin/coveralls",
